requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Brotli==1.0.9
dj-database-url==1.2.0
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
whitenoise==6.3.0
dj-database-url==1.2.0
django-cors-headers==3.14.0
django-rest-knox==4.2.0
drf-writable-nested==0.7.0
ipykernel==6.15.1
pip-chill==1.0.1
pycryptodome==3.15.0
pycryptodomex==3.15.0
pydes==2.0.1
virtualenv==20.19.0
asgiref==3.6.0
asttokens==2.0.8
backcall==0.2.0
cffi==1.15.1
colorama==0.4.5
cryptography==39.0.2
debugpy==1.6.3
decorator==5.1.1
distlib==0.3.6
dj-database-url==1.2.0
Django==4.1.7
django-cors-headers==3.14.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-writable-nested==0.7.0
entrypoints==0.4
executing==0.10.0
filelock==3.9.0
ipykernel==6.15.1
ipython==8.4.0
jedi==0.18.1
jupyter-client==7.3.4
jupyter-core==4.11.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
pip-chill==1.0.1
platformdirs==3.0.0
prompt-toolkit==3.0.30
psutil==5.9.1
pure-eval==0.2.2
pycparser==2.21
pycryptodome==3.15.0
pycryptodomex==3.15.0
pyDes==2.0.1
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==304
pyzmq==23.2.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.4.0
tornado==6.2
traitlets==5.3.0
tzdata==2022.7
virtualenv==20.19.0
wcwidth==0.2.5
asgiref==3.6.0
autopep8==2.0.1
Brotli==1.0.9
cffi==1.15.1
cryptography==39.0.2
dj-database-url==1.2.0
Django==4.1.7
django-cors-headers==3.14.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
whitenoise==6.3.0
asgiref==3.6.0
autopep8==2.0.1
Brotli==1.0.9
cffi==1.15.1
cryptography==39.0.2
dj-database-url==1.2.0
Django==4.1.7
django-cors-headers==3.14.0
django-multiselectfield==0.1.12
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
whitenoise==6.3.0
asgiref==3.6.0
autopep8==2.0.1
Brotli==1.0.9
cffi==1.15.1
cryptography==39.0.2
dj-database-url==1.2.0
Django==4.1.7
django-cors-headers==3.14.0
django-multiselectfield==0.1.12
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
whitenoise==6.3.0
